Switching ROCm TF to use the new hipfft library (instead of rocfft).
For ROCm 4.0 and prior, hipfft was a header-only "library" that was shipped along with the rocfft library. ROCm TF code uses the hipfft APIs defined in the "hipfft.h" header file, which was part of the package for rocfft library. ROCm TF would dynamically load the "rocfft" library at runtime.

From ROCm 4.1 onwards, it seems hipfft is a separate library from rocfft, and ROCm TF needs to use the "hipfft" library insteaf of the "rocfft" library. Specifcally, ROCm TF needs to
* pick up the "hipfft.h" header from the hipfft install area (instead of the rocfft install area).
* dynamically load the the "hipfft" library instead of the "rocfft" library
This commit has the two above changes

This change is required because without this change, the unit-tests that called the hipfft API ( `hipfftSetWorkArea` ) were failing when run with ROCm 4.1 (and without the changes in this commit).

ROCm TF continues to build (with ROCm 4.1) without the changes in this commit, because the old "hipfft.h" file in the rocfft install area is still present (do not know whether that is by design or accident)

unit-tests known to call the `hipfftSetWorkArea` API

//tensorflow/python/kernel_tests/linalg:linear_operator_circulant_test_gpu
//tensorflow/python/kernel_tests/signal:dct_ops_test
//tensorflow/python/kernel_tests/signal:mel_ops_test
//tensorflow/python/kernel_tests/signal:mfcc_ops_test
//tensorflow/python/kernel_tests/signal:spectral_ops_test
//tensorflow/python/ops/parallel_for:control_flow_ops_test_gpu