Updating the mapping of GCN Arch Name tokens to feature strings.
Prior to this commit, the GCN Arch Name "gfx908:+sramecc", would get mapped to
* { target : "gfx908", feature_string : "+sram-ecc"}

This commit changes the mapping to the following
* { target : "gfx908", feature_string : "+sramecc"}

(Notice the dash "-" is gone from "+sram-ecc")

The LLVM version being picked up by TF in this branch ( `develop-upstream-QA-rocm42` ) seems have changes withing it (as compared to the version being picked up in the `develop-upstream-QA-rocm41` branch), which require this change in the mapping.

This mapping is expected to be in a state of flux until all the AMDGPU target-id related changes have been pushed out to the upstream LLVM repo, and TF LLVM pointer is updated to pick those changes.