Revert "Have tf.nn.relu of NaN output NaN."
Adding NaN propagation to Relu, causing a huge perf degradation in the Eigen generated kernels for ROCm. This issue is being root-caused and will require hip-clang compiler changes to fix. Those fixes will come in a future ROCm release.

For the time being we need to revert this change to get the performance back.

This reverts commit 14dae44e4050ea3e44d1ca7124fd99899eb24879.