Fixing a bug in Conv3D implementation.
The following subtests were failing in the following container

tests
```
//bazel_pip/tensorflow/python/kernel_tests:conv_ops_3d_test              FAILED in 3 out of 3 in 121.6s
//bazel_pip/tensorflow/python/kernel_tests:conv_ops_3d_test_gpu          FAILED in 3 out of 3 in 43.2s
```

container - `compute-artifactory.amd.com:5000/rocm-plus-docker/framework/compute-rocm-dkms-no-npi-hipclang-tf-manylinux-env:6531_centos8_tf_r1.15-rocm-enhanced`

The reason for the failure was an invalid cache hit in the AutoTunerConv*. Some of the subtests within the above tests run the same convolution config first in float32 mode and then in float16 mode. The input tensor datatype was getting reported as having the same (incorrect) value in those two cases, and since that is the only thing different in the two configs, them being the same was resulting in an invalid cache hit for float16 case (when run after the flaot32 case). This resulted in the call to "Find" getting skipped, which then lead to the following error

```
MIOpen Error: /root/driver/MLOpen/src/ocl/convolutionocl.cpp:2536: No invoker was registered for convolution backward. Was find executed?
2021-05-12 01:58:05.452114: E tensorflow/stream_executor/dnn.cc:588] Failed to euqueue convolution on stream: miopenStatusUnknownError
```

This fix uses the dtype from the template arg (which should always be correct) instead of from the input tensor. This removes the cause of the invalid AutoTunerConv* cache hit, which results in "Find" getting called as expected, which makes the testcase pass.

The reason as to why the input tensor type is incorrect still needs to be investigated, skipping that for now since we have not seen this issue in newer TF releases.