Enable the use of rocsolver_Xpotrm_batched for the cholesky op in XLA.
The CUDA implementation uses a custom precompiled PTX kernel to map the a-data into an array pointers that can be consumed by the batched potrm op.

This will (obviously) not work for ROCm.

Instead, create a buffer on the host with the mapped pointers and copy this to the device prior to calling the batched potrm op.