Removing the nightly flag from the command to build the TF whl
This commit is for the `develop-upstream-QA-rocm51` which uses a static commit for building TF whls. Using the nightly flag results in the nightly versions of the dependent packages (like keras, tensorboard, tensorflow-estimator, etc) to get pulled in. The nightly versions are both unstable and at this point mnewer than static TF commit in the `develop-upstream-QA-rocm51` branch and hence potentially incompatible.

Removing the nightly flag will cause the TF whl to pull in the stable release versions of those dependent packages (instead of the unstable nightly ones), whihch is what we want