Hardcoding the `keras`, `tensorboard` and `tensoflow-estimator` package versions to 220306 nightly build
This is for the `develop-upstream-QA-rocm52` which was created on 20220307 (from the `develop-upstream` branch)

TF packages built from the `develop-upstream` branch are `nightly`, and have a dependency on the corresponding `nightly` builds of `keras`, `tensorboard` and `tensoflow-estimator`. This is okay for `develop-upstream` branch since it is sync'd with the `upstream/master` branch on a weekly basis.

However since the `develop-upstream-QA-rocm52` branch is static (i.e. it will not get any more commits from `upstream/master`), we need to make its dependencies static too. That is why this commit harcodes the `keras`, `tensorboard` and `tensoflow-estimator` package versions to 220306 nightly build, which is the last nightly build (for those packages) before the `develop-upstream-QA-rocm52` branch was created (from the `develop-upstream` branch)