Updating to `find_rocm_config.py` to look at newer heder locations first.
With a recent ROCm5.2 build, TF build started failing with the following error

```
...
ERROR: An error occurred during the fetch of repository 'local_config_rocm':
   Traceback (most recent call last):
	File "/root/tensorflow-upstream/third_party/gpus/rocm_configure.bzl", line 840, column 38, in _rocm_autoconf_impl
		_create_local_rocm_repository(repository_ctx)
	File "/root/tensorflow-upstream/third_party/gpus/rocm_configure.bzl", line 544, column 35, in _create_local_rocm_repository
		rocm_config = _get_rocm_config(repository_ctx, bash_bin, find_rocm_config_script)
	File "/root/tensorflow-upstream/third_party/gpus/rocm_configure.bzl", line 394, column 30, in _get_rocm_config
		config = find_rocm_config(repository_ctx, find_rocm_config_script)
	File "/root/tensorflow-upstream/third_party/gpus/rocm_configure.bzl", line 372, column 41, in find_rocm_config
		exec_result = _exec_find_rocm_config(repository_ctx, script_path)
	File "/root/tensorflow-upstream/third_party/gpus/rocm_configure.bzl", line 368, column 19, in _exec_find_rocm_config
		return execute(repository_ctx, [python_bin, "-c", decompress_and_execute_cmd])
	File "/root/tensorflow-upstream/third_party/remote_config/common.bzl", line 230, column 13, in execute
		fail(
Error in fail: Repository command failed
ERROR: #define "ROCRAND_VERSION" is either
  not present in file /opt/rocm-5.2.0-159/rocrand/include/rocrand_version.h OR
  its value is not an integer literal
...
```

For rocrand header files, TF looks at two separate locations (because location of header files changed in ROCm 5.1 and TF needs to be able to build with older ROCm versions too)

The (incorrect) assumption that TF makes is that the header files will only be present in either the newer location or the older location, but not both. So when it finds a file, it expects it to have a `#define` for the version number. This assumption is no longer true, since ROCm 5.2 builds now have a rocrand header file at the older location too. The header file at the older pulls in the header file from the newer location, but does not itself have a `#define` for the version number, and hence the error above.

This commit simply changes the order in which locations are looked up (newer first, after this change)