Updating `//tensorflow/python/keras/utils:multi_gpu_utils_test_2gpu` to account for nodes with >8 GPUs
The subtest `test_multi_gpu_test_invalid_devices` within `//tensorflow/python/keras/utils:multi_gpu_utils_test_2gpu` verifies that invalid GPU device ids are detected as such and an error message is issued for them. In one scenario, that subtest makes an implicit assumption that GPU device id `8` is an invalid device id. That assumption does not hold true for nodes which hand >8 GPUs, and hence the test fails on such nodes.

This commit updates that subtest to explicitly determine the invalid device id based on the number os GPUs on the node.