Updating `rocm_configure.bzl` to account for rocblas install changes in ROCM 5.2
ROCm 5.2 staging build 168 started failing TF build because of changes in the rocblas install

There is now a `$ROCM_PATH/include/rocblas` directory that did not exist before. This is expected as the old location `$ROCM_PATH/rocblas/include` is being deprecated. However this causes the rocblas headers to be copied over twice when creating `local_config_rocm`, once from `$ROCM_PATH/include` and then again from `$ROCM_PATH/rocblas/include`. So we need to stop the explicit copying of the headers from the `$ROCM_PATH/rocblas/include` directory (which is being deprecated)

Also it seems that the `$ROCM_PATH/rocblas/lib` directory is going away (atleast it is not present in the staging build mentioned above), and rocblas libraries noe get installed in `$ROCM_PATH/lib`. So updating `rocm_configure.bzl` to point to `$ROCM_PATH/lib` when looking for the rocblas libraries