Temporary workaround to fix build failure.
Ideally we should be including rocsolver header from the path `rocm/include/rocsolver/rocsolver.h` for ROCm version 5.2 and newer, which is what the previous commit to this file does. Controlling the header file path based on ROCm version requires pulling in the `rocm/rocm_config.h` header (which defines the TF_ROCM_VERSION macro). Unfortunately pulling in that header, triggers compile failures because it enbales code-paths we thought we were testing, but turns out we were not.

So reverting to using the `rocm/include/rocsolver.h` path for including the rocsolver header. That path will work for ROCm 5.0 and above, but will be deprecated eventually

This commit needs to be reverted once the build failures, triggered by including the `rocm/rocm_config.h` file in the `rocsolver_wrapper.h` header file, are resolved.