Removing the setting of `ROCBLAS_TENSILE_LIBPATH` in `configure.py`
With ROCm 5.2, TF unit tests started erroring out when loading rocblas because of a change in the location of rocblas tensile libraries

old location : `$ROCM_PATH/lib/library`
new location : `$ROCM_PATH/lib/rocblas/library`

It seems that we no longer need to explicitly set the `ROCBLAS_TENSILE_LIBPATH` env var. When not set, rocblas will look for it in the location determined by where the rocblas library itself was loaded. That should work for all ROCm 5.0, 5.1, 5.2 and beyond.

Therefore this commit removes the setting of `ROCBLAS_TENSILE_LIBPATH` in `configure.py`